requirements.txt
streamlit
pandas
python-dotenv
requests
sqlparse
PyYAML
